Rename optuna.testing.integration with optuna.testing.pruner
