+++ tests/integration_tests/test_pytorch_ignite.py
@@ -5,7 +5,7 @@ from ignite.engine import Engine
 import pytest
 
 import optuna
-from optuna.testing.integration import DeterministicPruner
+from optuna.testing.pruner import DeterministicPruner
 
 
 def test_pytorch_ignite_pruning_handler() -> None:
